Arreglando flujo de juego
