+++ src/index.html
@@ -6,7 +6,9 @@
     <title>Document</title>
   </head>
   <body>
-    <div class="contenedor" id="contenedor"></div>
+    <div class="container" id="container">
+      <div class="container-bienvenida" id="container-bienvenida"></div>
+    </div>
 
     <div id="selectContainer"></div>
     <div id="questionContainer"></div>
